# docs/swagger.yaml
definitions:
...
  helpers.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.TicketInventory:
    properties:
      event_session_id:
        type: string
      held_count:
        type: integer
      id:
        type: string
      sold_count:
        type: integer
      ticket_tier_id:
        type: string
      total_quota:
        type: integer
      updated_at:
        type: string
      version:
        type: integer
    type: object
  models.TicketTier:
    properties:
      created_at:
        type: string
      event_id:
        type: string
      id:
        type: string
      is_seated:
        type: boolean
      max_per_transaction:
        type: integer
      name:
        type: string
      price:
        type: number
      sale_end_at:
        type: string
      sale_start_at:
        type: string
      tier_type:
        type: string
      updated_at:
        type: string
      venue_section_id:
        type: string
    type: object
...
  ticket.CreateInventoryRequest:
    properties:
      event_session_id:
        type: string
      total_quota:
        minimum: 0
        type: integer
    required:
    - event_session_id
    type: object
  ticket.CreateTicketTierRequest:
    properties:
      is_seated:
        type: boolean
      max_per_transaction:
        type: integer
      name:
        maxLength: 100
        type: string
      price:
        minimum: 0
        type: number
      sale_end_at:
        type: string
      sale_start_at:
        type: string
      tier_type:
        enum:
        - early_bird
        - regular
        - vip
        - vvip
        - group_package
        type: string
      venue_section_id:
        type: string
    required:
    - name
    - tier_type
    type: object
  ticket.UpdateTicketTierRequest:
    properties:
      is_seated:
        type: boolean
      max_per_transaction:
        type: integer
      name:
        maxLength: 100
        type: string
      price:
        minimum: 0
        type: number
      sale_end_at:
        type: string
      sale_start_at:
        type: string
      tier_type:
        enum:
        - early_bird
        - regular
        - vip
        - vvip
        - group_package
        type: string
      venue_section_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API for the Konsera ticketing platform
  title: Konsera API
  version: "1.0"
paths:
...
  /events/{event_id}/ticket-tiers:
    get:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.TicketTier'
            type: array
      security:
      - BearerAuth: []
      summary: List ticket tiers
      tags:
      - Ticket Tiers
    post:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Ticket tier
        in: body
        name: tier
        required: true
        schema:
          $ref: '#/definitions/ticket.CreateTicketTierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TicketTier'
      security:
      - BearerAuth: []
      summary: Create ticket tier
      tags:
      - Ticket Tiers
  /events/{event_id}/ticket-tiers/{tier_id}:
    delete:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Ticket tier ID
        in: path
        name: tier_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/helpers.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete ticket tier
      tags:
      - Ticket Tiers
    get:
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Ticket tier ID
        in: path
        name: tier_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TicketTier'
      security:
      - BearerAuth: []
      summary: Get ticket tier
      tags:
      - Ticket Tiers
    put:
      consumes:
      - application/json
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - description: Ticket tier ID
        in: path
        name: tier_id
        required: true
        type: string
      - description: Ticket tier changes
        in: body
        name: tier
        required: true
        schema:
          $ref: '#/definitions/ticket.UpdateTicketTierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TicketTier'
      security:
      - BearerAuth: []
      summary: Update ticket tier
      tags:
      - Ticket Tiers
...
  /ticket-tiers/{tier_id}/inventory:
    post:
      consumes:
      - application/json
      parameters:
      - description: Ticket tier ID
        in: path
        name: tier_id
        required: true
        type: string
      - description: Inventory
        in: body
        name: inventory
        required: true
        schema:
          $ref: '#/definitions/ticket.CreateInventoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TicketInventory'
      security:
      - BearerAuth: []
      summary: Create inventory
      tags:
      - Ticket Inventories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"